Move all Rust<->OCaml conversion code to tezos/api.
